+++ r-package/R/read_census_tract.R
@@ -70,7 +70,7 @@ read_census_tract <- function(code_tract, year=2010, zone = "urban", simplified=
       file_url <- as.character(temp_meta$download_path)
 
       # download files
-      temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+      temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
       # check if download failed
       if (is.null(temp_sf)) { return(invisible(NULL)) }
@@ -105,7 +105,7 @@ read_census_tract <- function(code_tract, year=2010, zone = "urban", simplified=
 
         }
       # download files
-      sf <- download_gpkg(file_url, progress_bar = showProgress)
+      sf <- download_gpkg(file_url, showProgress = showProgress)
 
       # check if download failed
       if (is.null(sf)) { return(invisible(NULL)) }