+++ r-package/R/read_disaster_risk_area.R
@@ -33,7 +33,7 @@ read_disaster_risk_area <- function(year=2010, simplified=TRUE, showProgress=TRU
   file_url <- as.character(temp_meta$download_path)
 
   # download files
-  temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+  temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
   # check if download failed
   if (is.null(temp_sf)) { return(invisible(NULL)) }