+++ r-package/R/read_health_region.R
@@ -41,7 +41,7 @@ read_health_region <- function(year=2013, macro=FALSE, simplified=TRUE, showProg
   file_url <- as.character(temp_meta$download_path)
 
   # download files
-  temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+  temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
   # check if download failed
   if (is.null(temp_sf)) { return(invisible(NULL)) }