+++ r-package/R/read_immediate_region.R
@@ -42,7 +42,7 @@ read_immediate_region <- function(code_immediate="all", year=2019, simplified=TR
   file_url <- as.character(temp_meta$download_path)
 
   # download files
-  temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+  temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
   # check if download failed
   if (is.null(temp_sf)) { return(invisible(NULL)) }