+++ r-package/R/read_intermediate_region.R
@@ -44,7 +44,7 @@ read_intermediate_region <- function(code_intermediate="all", year=2019, simplif
   file_url <- as.character(temp_meta$download_path)
 
   # download files
-  temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+  temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
   # check if download failed
   if (is.null(temp_sf)) { return(invisible(NULL)) }