+++ r-package/R/read_meso_region.R
@@ -50,7 +50,7 @@ read_meso_region <- function(code_meso = "all",
     file_url <- as.character(temp_meta$download_path)
 
     # download files
-    temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+    temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
     # check if download failed
     if (is.null(temp_sf)) { return(invisible(NULL)) }
@@ -71,7 +71,7 @@ read_meso_region <- function(code_meso = "all",
 
 
     # download files
-    temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+    temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
     # check if download failed
     if (is.null(temp_sf)) { return(invisible(NULL)) }