+++ r-package/R/read_pop_arrangements.R
@@ -28,7 +28,7 @@ read_pop_arrangements <- function(year=2015, simplified=TRUE, showProgress=TRUE)
   file_url <- as.character(temp_meta$download_path)
 
   # download files
-  temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+  temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
   # check if download failed
   if (is.null(temp_sf)) { return(invisible(NULL)) }