+++ r-package/R/read_statistical_grid.R
@@ -50,7 +50,7 @@ read_statistical_grid <- function(code_grid, year=2010, showProgress=TRUE){ # no
       file_url <- as.character(temp_meta$download_path)
 
       # download files
-      temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+      temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
       # check if download failed
       if (is.null(temp_sf)) { return(invisible(NULL)) }
@@ -84,7 +84,7 @@ read_statistical_grid <- function(code_grid, year=2010, showProgress=TRUE){ # no
       file_url <- as.character(subset(temp_meta, code %in% grid_ids)$download_path)
 
       # download gpkg
-      temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+      temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
       # check if download failed
       if (is.null(temp_sf)) { return(invisible(NULL)) }
@@ -102,7 +102,7 @@ read_statistical_grid <- function(code_grid, year=2010, showProgress=TRUE){ # no
     file_url <- as.character(subset(temp_meta, code== code_grid)$download_path)
 
     # download files
-    temp_sf <- download_gpkg(file_url, progress_bar = showProgress)
+    temp_sf <- download_gpkg(file_url, showProgress = showProgress)
 
     # check if download failed
     if (is.null(temp_sf)) { return(invisible(NULL)) }