+++ r-package/man/roxygen/templates/cache.R
@@ -1,0 +1,5 @@
+#' @param cache Logical. Whether the function should read the data cached
+#'        locally, which is faster. Defaults to `cache = TRUE`. By default,
+#'        `geobr` stores data files in a temporary directory that exists only
+#'        within each R session. If `cache = FALSE`, the function will download
+#'        the data again and overwrite the local file.